Add files via upload
fixes: advanced options work (limited testing) 

bugs: can't move and rename at the same time.  Think this has to do with some part in the 'try' block of moveFile(), but I'm not sure what's causing it

etc: lots of repetitive code.  if time allows, it would be best to break some of the advanced options conditions to their own class